Moved POJO classes into dto package
