+++ bundles/org.openhab.binding.openweathermap/src/main/java/org/openhab/binding/openweathermap/internal/dto/base/Rain.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.openweathermap.internal.model.base;
+package org.openhab.binding.openweathermap.internal.dto.base;
 
 import org.eclipse.jdt.annotation.Nullable;
