+++ bundles/org.openhab.binding.openweathermap/src/main/java/org/openhab/binding/openweathermap/internal/dto/forecast/daily/FeelsLikeTemp.java
@@ -10,17 +10,15 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.openweathermap.internal.model.forecast.daily;
+package org.openhab.binding.openweathermap.internal.dto.forecast.daily;
 
 /**
- * Generated Plain Old Java Objects class for {@link Temp} from JSON.
+ * Generated Plain Old Java Objects class for {@link FeelsLikeTemp} from JSON.
  *
  * @author Christoph Weitkamp - Initial contribution
  */
-public class Temp {
+public class FeelsLikeTemp {
     private Double day;
-    private Double min;
-    private Double max;
     private Double night;
     private Double eve;
     private Double morn;
@@ -31,22 +29,6 @@ public class Temp {
 
     public void setDay(Double day) {
         this.day = day;
-    }
-
-    public Double getMin() {
-        return min;
-    }
-
-    public void setMin(Double min) {
-        this.min = min;
-    }
-
-    public Double getMax() {
-        return max;
-    }
-
-    public void setMax(Double max) {
-        this.max = max;
     }
 
     public Double getNight() {
