+++ bundles/org.openhab.binding.openweathermap/src/main/java/org/openhab/binding/openweathermap/internal/dto/weather/Main.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.openweathermap.internal.model.forecast.hourly;
+package org.openhab.binding.openweathermap.internal.dto.weather;
 
 import org.eclipse.jdt.annotation.Nullable;
 
@@ -25,18 +25,12 @@ public class Main {
     private Double temp;
     @SerializedName("feels_like")
     private @Nullable Double feelsLikeTemp;
+    private Double pressure;
+    private Integer humidity;
     @SerializedName("temp_min")
     private Double tempMin;
     @SerializedName("temp_max")
     private Double tempMax;
-    private Double pressure;
-    @SerializedName("sea_level")
-    private Double seaLevel;
-    @SerializedName("grnd_level")
-    private Double grndLevel;
-    private Integer humidity;
-    @SerializedName("temp_kf")
-    private Integer tempKf;
 
     public Double getTemp() {
         return temp;
@@ -54,6 +48,22 @@ public class Main {
         this.feelsLikeTemp = feelsLikeTemp;
     }
 
+    public Double getPressure() {
+        return pressure;
+    }
+
+    public void setPressure(Double pressure) {
+        this.pressure = pressure;
+    }
+
+    public Integer getHumidity() {
+        return humidity;
+    }
+
+    public void setHumidity(Integer humidity) {
+        this.humidity = humidity;
+    }
+
     public Double getTempMin() {
         return tempMin;
     }
@@ -68,45 +78,5 @@ public class Main {
 
     public void setTempMax(Double tempMax) {
         this.tempMax = tempMax;
-    }
-
-    public Double getPressure() {
-        return pressure;
-    }
-
-    public void setPressure(Double pressure) {
-        this.pressure = pressure;
-    }
-
-    public Double getSeaLevel() {
-        return seaLevel;
-    }
-
-    public void setSeaLevel(Double seaLevel) {
-        this.seaLevel = seaLevel;
-    }
-
-    public Double getGrndLevel() {
-        return grndLevel;
-    }
-
-    public void setGrndLevel(Double grndLevel) {
-        this.grndLevel = grndLevel;
-    }
-
-    public Integer getHumidity() {
-        return humidity;
-    }
-
-    public void setHumidity(Integer humidity) {
-        this.humidity = humidity;
-    }
-
-    public Integer getTempKf() {
-        return tempKf;
-    }
-
-    public void setTempKf(Integer tempKf) {
-        this.tempKf = tempKf;
     }
 }
